use Menhir's support for error messages
The way the error messages are generated is by asking Menhir to
produce a description of the possible error modes of the grammar in
a file parser.messages (from parser.mly) with placeholder for error
messages. Then I fill the placeholder with Gabriel-generater error
messages, based on Menhir's explanations of the error situation
(in comments). Then I can ask Menhir to generate the file
parser_messages.ml who maps an error state (of the parsing automaton)
to the right message.

Finally, a rather tricky bit of Menhir machinery needs to be done to
obtain the error state in case of a parsing error. This goes in
parse.ml, and it is done by carefully reading the Menhir manual.