explicit scope
construct the scoping rules in such a way that the proof
yields the correct scope for every label.

this greatly simplifies the checking of osr since we can just
check that the osr declares the whole scope.

to be able to ground the beginning of the program with the
empty scope we define the first label to be the start label.