make scope strict: join points must have exactly the same variables
Not many places in the testsuite needed to be changed, which is a good
surprise.

The case of

    const x = e
    branch x l1 l2

is a bit annoying, as makes backward jumps more painful: one cannot
drop x before jumping ("branch" needs x in scope), one cannot jump to
a label before in the program (x needs to be dropped first), so one
has to have a local label below to drop x first and only then jump
backward:

  loop:
    const x = ...
    branch x loop end
  end:

needs to be rewritten into

  loop:
    const x = ...
    branch x drop_and_loop end
  drop_and_loop:
    drop x
    goto loop
  end:

There is currently no decent error message in case of
a conflict. Providing nice errors requires a bit more work in the
analysis, to keep tracks of where the inferred scopes are coming from
(currently we don't have enough context at the
exception-raising site).