allow sometimes uninitialized variables
the updated scope analysis allows variables to be
inconsistently initialized/uninitialized on controlflow
merge. after discussions with @gasche it's not sure
if this is what we want in the long term. But for now
I'd like to proceed with it since it makes code motion
easier.

The new analysis is still conservative as in: it's not
possible to use an uninitialized variable.