make annotations part of the version
a version is now a

    { label; instrs; annotations; }

instead of keeping a separate list of instructions and annotations
in the function.

Note: annotations is an option since transformation do not keep
the annotations. we could also put an array full of None there
(ie. no annotation at every instruction) but that seems a bit
wasteful.