remove literals from instr
There was some confusion over the usage of the term 'literal' in
instr. Since all variants of 'literal' where used to describe runtime
values I changed it to 'tagged_value'. Consequently a simple
expression now contains a 'Constant of tagged_value'.

This change conflates the old value and literal type. Unfortunately
this forces us to have `Fun_ref of string` instead of the more
natural `Fun_ref of afunction`. If we wanted the the later we would
need some intermediate parsetree or a preparser to build up the
table of functions.