fix minimize_lifetime
The way we tried to "saturate" lifetime information from a backward
pass result was incorrect on programs with multi-predecessor
multi-successor instructions -- for the same reason that we had to
split edges in the drop hoisting code.

On

      var a = (10 + 0)
      var b = 20
      var c = 30
      branch t la lb
    la:
      branch t l1 l2
    lb:
      branch t l2 l3
    l1:
      print (a + 0)
      goto fin
    l2:
      print b
      goto fin
    l3:
      print c
      goto fin
    fin:
      stop 0

for example, it would generate the following program

     var a = (10 + 0)
     var b = 20
     var c = 30
     branch t la lb
    la:
     drop c
     branch t l1 l2
    lb:
     drop a
     branch t l2 l3
    l1:
     drop b
     print (a + 0)
     drop a
     goto fin
    l2:
     drop a
     drop c
     print b
     drop b
     goto fin
    l3:
     drop b
     print c
     drop c
     goto fin
    fin:
     stop 0

which does not respect our scoping rules (`la` and `lb` may merge in
`l2` yet have distinct scopes).

I rewrote the pass to use the drop-hoisting code instead. First,
insert drop maximally (this is simplified by inserting
returns maximally), then hoist all the drops. This may split edges.