make transform_hoist more robust
Currently transform_hoist assumes that push a drop above will never
encounter a `stop` or `return` instruction. Indeed, even an
instruction that would be immediately after such an instruction would
not have it in its predecessors. But the current implementation
hardcodes "look at pc-1" when pushing over a label, which means that
the assert fails in code such as:

    stop 0
    goto label # unique jump to this label
    label:
    drop x

The code will hoist `drop x` above `goto label`, and immediately try
to push it above `stop 0` (as `pc-1` relatively to the
`goto` instruction), which raises the assertion failure.

I changed the transform_hoist code to not rely on fixed pc addressing,
but instead transform all predecessors. I also removed the separate
remove_fallthrough_to_label pass which somewhat obscured the issue and
became half-redundant with the new, more flexible implementation, and
consequently had to strengthen Edit.split_edge which assumed the
absence of fallthrough labels.

My moral for this story would be: unless our representation *enforces*
strong invariant, it's better to write code that is flexible in
absence of these invariants than to rely on preconditions.