combine checkpoint label with osr instruction
Checkpoint labels are a bit of a hassle when it comes to transformations.
They have to be threated differently from the other labels. For example
they have to be preserved even if there is no goto/branch using them.

Also if we want to enforce some structure on the cfg in the future they
will present an exception again.

Therefore I combined them into the osr instruction.