don't wrap exceptions
the wrapping of the exception makes it harder to trace back to
the point of origin of the exception. Since all those exceptions
in the optimizer should be fixed it does not make much sense
to have it there. When I want to debug a problem in an optimizer
I always have to remove this code first to be able to see the
origin of the exn.