inlining: fix var and label renaming
To rename variables and labels we cannot call `fresh` on all
existing variables in parallel. For example consider the function:

```
caller:
  var x = 1

callee:
  var x   = 1
  var x_1 = 2
```

In general any try to do the renaming in parallel makes unreasonable
assumptions about how `fresh` is implemented. For example it would be
perfectly reasonable for `fresh` to always return `foo` if it's fresh.

Thus the only way of using fresh which does not make assumptions on
how it works is to always feed back the generated names to avoid generating
the same name again.