use the OCaml native compiler
The bytecode compiler compiles faster (and is more incremental) than
the native compiler, so I guess I used it at the start for convenience
of development. Given that build times remain small and that we are
starting to hit performance issues, let's move to the real deal now --
as we probably should have a while ago.