extend the syntax of osr instruction to create additional frames
this allows osr to create additional frames. The syntax is:

osr l [...] (...) [...], (f,v,l)[var x = $, var y = 2], (...)[...]

There are several parts to it.

* The label l
* A list of conditions, eg. [x==2, x==3]
* The osr target, eg. (f,v,l)
* The varmap definition of the top frame, eg. [var x = 2]
* Finally (new) a comma separated list of extra frames.

An extra frame has the syntax `(f,v,l) [var x = $, ...]`
* The `()` part specifies return target of the additional frame
* The `[]` specifies the extra environment to create.
  It is similar to the normal varmap but needs to begin with an
  additional `var x = $` to specify which variable will hold the
  function result.

For example
   var x = 1
   call y = foo()
  after:

can be simulated by an osr extra frame:
  (f,v,after) [var y = $, var x = 1]

Each of those extra frame entries defines one continuation. Due to the
way we create continuations we have to in include the result variable
in the varmap with the special `$` syntax.