change transform_inline to be aware of checkpoints
To properly support osr across inlining the inlining pass
needs to know if there is a checkpoint after the function
call at the caller site.

This patch refactors compute_inlining_order to not just
return the order, but in fact an annotated graph that
contains all the information for the inliner to go forward.

At the core there is the `inlining_candidate` which contains
all information to apply one inlining. The inlinings are
performed recursively, all in one go.

At the end a new version is returned which can be added to
the first caller function.

Currently inlinings are skipped when changes to the osr
points would be needed.