extend transform_inline to handle osr
when creating inline_candidates we keep track of the safepoints
around the function call. We use the existing osr map to create
an extended version.

For example given:

  function caller
  version active
    var x = 1
    call y = bla (...)
    osr after_call ...  [var x = x]

We create the following frame map:

    (caller, active, after_call) [var y = $, var x = x]

This can be used to fixup the osrs in the inlinee. For example

  function bla
  version active
    osr cp_1 [] (bla, base, cp_1) [...]

Can be inlined as

    var x = 1
    osr cp_1 [] (bla, base, cp_1) [...], (caller, active, after_call) [var y = $, var x = x]
    osr after_call ...  [var x = x]