fix hoist_drop
there was a very subtle bug which could cause an unrelated instruction
to be pulled. In the case:

 branch x $a $b
$a:
 drop y
 drop x
$b:
 drop y
 drop x

the Need_pull could end up twice on the work list which causes both
drop x and drop y to be pulled. The solution is to only add the
branch once to the worklist