conservative fix_scope
I am not super confident in the correctnes of this generic fix_scope
pass.

But I noticed that inserting random variables into scope can only
break the scope at merge points. There needs to be a goto right before
the scope is broken. Inserting before goto is quite painless (since the
goto itself does not interact with the scope). Thus the simple and
conservative option is to only fix the scope when necessary right
before a goto.