add annotations to inject assumptions
guard_hint is a way of injecting an assumption.

in the assumption pass sourir will promote guard_hint into the nearest
condition for an osr instruction.

The instruction is called guard and not osr_hint because I would like to
rename osr to guard too in the future. More on that later...