gen_next_version incorrectly copied the scope
the checkpoint copied by gen_next_version was only updated
to have the correct target, but not the correct target
scope.

When we call gen_next_version we need to restore the invariant
of add_empty_osr, namely that the scope captured by
the checkpoint is equal to the local scope.

also the extra frames can be dumped since we chain checkpoints
and the checkpoint one level down will be responsible for
creating them.